refactor: centralize AI model operations
Keep both AI-spacing entry points focused on their high-level flows by moving session and LanguageModel details into model.ts.
