+++ docs/ai-spacing.md
@@ -66,7 +66,7 @@ Labels zip against the candidates by index, with `null` for a failed candidate. 
 
 ### Classifier and service worker (`browser-extensions/chrome`)
 
-Each ambiguous shape's `PromptSpec` carries its shipping prompt bytes in an extension-local constant file, tagged with a `version`: the system prompt, the markless question template (an untouched sentence, with the question pointing at the symbol by quoting the character before it), the candidate labels, and the canonical menu order. `ai-spacing/in-service-worker.ts` holds the registry, keyed by `kind`, and keeps one base session per kind in a `Map<kind, Promise<LanguageModel>>`, since each shape's system prompt rides in its own session's `initialPrompts`. A session is created at `temperature 0` and `topK 1`, cloned per candidate through a sequential loop, and recreated when the MV3 worker wakes. What is cached is the promise rather than the session, so batches that arrive while one `create()` is in flight share it, and a rejected `create()` drops that kind's entry, because the model can arrive later through the options-page download. A message whose `kind` has no prompt spec registered answers `{ok: false}`, the same clean no the page gets for an absent model. A message with no candidates creates the base session and answers an empty list, which is all the warm-up needs.
+Each ambiguous shape's `PromptSpec` carries its shipping prompt bytes in an extension-local constant file, tagged with a `version`: the system prompt, the markless question template (an untouched sentence, with the question pointing at the symbol by quoting the character before it), the candidate labels, and the canonical menu order. `ai-spacing/in-service-worker.ts` holds the registry, keyed by `kind`, passes the selected spec and candidates to `model.ts`, and wraps the result in a message response. `model.ts` owns session creation, cloning, prompting, response validation, and cleanup. It keeps one base session per kind in a `Map<kind, Promise<LanguageModel>>`, since each shape's system prompt rides in its own session's `initialPrompts`. A session is created at `temperature 0` and `topK 1`, cloned per candidate through a sequential loop, and recreated when the MV3 worker wakes. What is cached is the promise rather than the session, so batches that arrive while one `create()` is in flight share it, and a rejected `create()` drops that kind's entry, because the model can arrive later through the options-page download. A message whose `kind` has no prompt spec registered answers `{ok: false}`, the same clean no the page gets for an absent model. A message with no candidates creates the base session and answers an empty list, which is all the warm-up needs.
 
 `v21-zh-english-labels` keeps the Chinese instructions and glosses from `v21-zh`, but uses `signed-number`, `range-or-separator`, and `unsure` directly in the menu and response constraint. The worker parses JSON and accepts only those labels, without translation. Model accuracy has not been re-evaluated for this variant.
